+++ src/index.test.js
@@ -8,7 +8,8 @@ jest.mock('DirectUploadProvider', () =>
   jest.fn(props => <div data-component="DirectUploadProvider" {...props} />)
 )
 
-jest.mock('csrfHeader', () => () => ({ 'X-CSRF-Token': 'qwertyuiop' }))
+const mockCSRFHeader = { 'X-CSRF-Token': 'qwertyuiop' }
+jest.mock('csrfHeader', () => () => mockCSRFHeader)
 
 const endpoint = {
   path: '/users/1',
@@ -68,5 +69,62 @@ describe('ActiveStorageProvider', () => {
   it('doesn’t hit the endpoint if handleUpload is called with no files', async () => {
     await tree.props.onSuccess([])
     expect(fetch).not.toHaveBeenCalled()
+  })
+
+  describe('if custom headers are provided', () => {
+    const customHeaders = { 'TEST-HEADER': 'testValue' }
+
+    beforeEach(() => {
+      component = renderComponent({ headers: customHeaders })
+      tree = component.toJSON()
+    })
+
+    it('hits the given endpoint with the custom headers', async () => {
+      await tree.props.onSuccess(['signedId'])
+      const headers = new Headers({
+        Accept: 'application/json',
+        'Content-Type': 'application/json',
+        ...mockCSRFHeader,
+        ...customHeaders,
+      })
+      expect(fetch).toHaveBeenCalledWith(
+        endpoint.path,
+        expect.objectContaining({
+          method: endpoint.method,
+          body: JSON.stringify({
+            [endpoint.model.toLowerCase()]: { avatar: 'signedId' },
+          }),
+          headers,
+        })
+      )
+    })
+
+    describe('if an X-CSRF-Token is provided', () => {
+      const customHeaders = { 'X-CSRF-Token': 'testToken' }
+
+      beforeEach(() => {
+        component = renderComponent({ headers: customHeaders })
+        tree = component.toJSON()
+      })
+
+      it('uses the provided X-CSRF-Token', async () => {
+        await tree.props.onSuccess(['signedId'])
+        const headers = new Headers({
+          Accept: 'application/json',
+          'Content-Type': 'application/json',
+          ...customHeaders,
+        })
+        expect(fetch).toHaveBeenCalledWith(
+          endpoint.path,
+          expect.objectContaining({
+            method: endpoint.method,
+            body: JSON.stringify({
+              [endpoint.model.toLowerCase()]: { avatar: 'signedId' },
+            }),
+            headers,
+          })
+        )
+      })
+    })
   })
 })